// src/types.ts
export enum ContentCategory {
    HEADER = "HEADER",
    FOOTER = "FOOTER",
    HERO = "HERO",
    ABOUT = "ABOUT",
    SERVICES = "SERVICES",
    TEXT = "TEXT",
    CONTACT = "CONTACT"
}

export interface ProjectStruct {
    id: string
    title: string
    pages: Record<string, PageStruct>
    colourScheme: ColourScheme
    fontScheme: FontScheme
}

export interface PageStruct {
    id: string
    path: string
    description: string
    content: ContentStruct[]
    contentActive: ContentStruct | null
    colourScheme: ColourScheme
    fontScheme: FontScheme
}

export interface ContentStruct {
    id: string
    description: string
    Component: React.FC | string
    categories: ContentCategory[]
    props: any
}